libc: Change getlogin_r()'s second argument to size_t, as POSIX likes it.
https://pubs.opengroup.org/onlinepubs/9699919799/functions/getlogin.html

Also adjust the getconf syscall, aka _getconf() in userland, which is
called by getconf(3) and getconf_r(3).

Approved-by: dillon
Tested-by:   zrj

+++ sys/sys/sysproto.h
@@ -326,7 +326,7 @@ struct	getlogin_args {
 	struct sysmsg sysmsg;
 #endif
 	char *	namebuf;	char namebuf_[PAD_(char *)];
-	u_int	namelen;	char namelen_[PAD_(u_int)];
+	size_t	namelen;	char namelen_[PAD_(size_t)];
 };
 struct	setlogin_args {
 #ifdef _KERNEL
